FMOD Studio dlls will always be copied (even if not used)
Our wrapper is a massive dynamic library around both the Core and Studio DLLs so both dependencies need to exist for the library to load
To change this behaviour FMOD would need to be compiled into 2 different dlls the YYFMODCore.dll and YYFMODStudio.dll which is not the case.
The IDE right now doesn't support conditional compilation so that's how it will work for the time being.